# backend-bk/internal/handler/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Collection Game API",
    "version": "v1",
    "description": "OpenAPI spec for Collection Game backend"
  },
  "servers": [
    {
      "url": "/",
      "description": "Root server"
    }
  ],
  "tags": [
    { "name": "Health" },
    { "name": "Auth" },
    { "name": "User" },
    { "name": "Characters" },
    { "name": "Summon" },
    { "name": "Dungeons" },
    { "name": "Arena" },
    { "name": "Quests" },
    { "name": "Guilds" },
    { "name": "Shop" }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAuth" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAuth" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh token",
        "responses": {
          "200": {
            "description": "Refreshed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAuthToken" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseMessage" }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/profile": {
      "get": {
        "tags": ["User"],
        "summary": "Get profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseUser" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Update profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseUser" }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/inventory": {
      "get": {
        "tags": ["User"],
        "summary": "Get inventory",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Inventory",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters": {
      "get": {
        "tags": ["Characters"],
        "summary": "Get user characters",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Characters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{id}": {
      "get": {
        "tags": ["Characters"],
        "summary": "Get character detail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Character detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{id}/level-up": {
      "post": {
        "tags": ["Characters"],
        "summary": "Level up character",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LevelUpRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated character",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{id}/awaken": {
      "post": {
        "tags": ["Characters"],
        "summary": "Awaken character",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Updated character",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/summon/normal": {
      "post": {
        "tags": ["Summon"],
        "summary": "Normal summon",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Summon result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/summon/premium": {
      "post": {
        "tags": ["Summon"],
        "summary": "Premium summon",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Summon result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/summon/rates": {
      "get": {
        "tags": ["Summon"],
        "summary": "Get summon rates",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Rates",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseObject" }
              }
            }
          }
        }
      }
    },
    "/api/v1/dungeons": {
      "get": {
        "tags": ["Dungeons"],
        "summary": "Get dungeons",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Dungeons" } }
      }
    },
    "/api/v1/dungeons/{id}": {
      "get": {
        "tags": ["Dungeons"],
        "summary": "Get dungeon detail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Dungeon detail" } }
      }
    },
    "/api/v1/dungeons/{id}/enter": {
      "post": {
        "tags": ["Dungeons"],
        "summary": "Enter dungeon",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Entered" } }
      }
    },
    "/api/v1/dungeons/{id}/complete": {
      "post": {
        "tags": ["Dungeons"],
        "summary": "Complete dungeon",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Completed" } }
      }
    },
    "/api/v1/arena/ranking": {
      "get": {
        "tags": ["Arena"],
        "summary": "Get arena ranking",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Ranking" } }
      }
    },
    "/api/v1/arena/attack": {
      "post": {
        "tags": ["Arena"],
        "summary": "Attack",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Attack result" } }
      }
    },
    "/api/v1/arena/history": {
      "get": {
        "tags": ["Arena"],
        "summary": "Get arena history",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "History" } }
      }
    },
    "/api/v1/quests/daily": {
      "get": {
        "tags": ["Quests"],
        "summary": "Get daily quests",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Daily quests" } }
      }
    },
    "/api/v1/quests/weekly": {
      "get": {
        "tags": ["Quests"],
        "summary": "Get weekly quests",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Weekly quests" } }
      }
    },
    "/api/v1/quests/achievements": {
      "get": {
        "tags": ["Quests"],
        "summary": "Get achievements",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Achievements" } }
      }
    },
    "/api/v1/quests/{id}/claim": {
      "post": {
        "tags": ["Quests"],
        "summary": "Claim quest",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Claimed" } }
      }
    },
    "/api/v1/guilds": {
      "get": {
        "tags": ["Guilds"],
        "summary": "Get guilds",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Guilds" } }
      },
      "post": {
        "tags": ["Guilds"],
        "summary": "Create guild",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Created" } }
      }
    },
    "/api/v1/guilds/{id}": {
      "get": {
        "tags": ["Guilds"],
        "summary": "Get guild",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Guild" } }
      },
      "put": {
        "tags": ["Guilds"],
        "summary": "Update guild",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Updated" } }
      }
    },
    "/api/v1/guilds/{id}/join": {
      "post": {
        "tags": ["Guilds"],
        "summary": "Join guild",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Joined" } }
      }
    },
    "/api/v1/guilds/{id}/leave": {
      "post": {
        "tags": ["Guilds"],
        "summary": "Leave guild",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Left" } }
      }
    },
    "/api/v1/guilds/{id}/members": {
      "get": {
        "tags": ["Guilds"],
        "summary": "Get guild members",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
        "responses": { "200": { "description": "Members" } }
      }
    },
    "/api/v1/guilds/my": {
      "get": {
        "tags": ["Guilds"],
        "summary": "Get my guild",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "My guild" } }
      }
    },
    "/api/v1/shop/items": {
      "get": {
        "tags": ["Shop"],
        "summary": "Get shop items",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Items" } }
      }
    },
    "/api/v1/shop/purchase": {
      "post": {
        "tags": ["Shop"],
        "summary": "Purchase item",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Purchased" } }
      }
    },
    "/api/v1/shop/history": {
      "get": {
        "tags": ["Shop"],
        "summary": "Get purchase history",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "History" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "schemas": {
      "ApiResponseAuth": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/AuthResponse" },
          "error": { "type": ["string", "null"] },
          "timestamp": { "type": "string" }
        }
      },
      "ApiResponseAuthToken": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/AuthTokenResponse" },
          "error": { "type": ["string", "null"] },
          "timestamp": { "type": "string" }
        }
      },
      "ApiResponseUser": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/User" },
          "error": { "type": ["string", "null"] },
          "timestamp": { "type": "string" }
        }
      },
      "ApiResponseMessage": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "object" },
          "error": { "type": ["string", "null"] },
          "timestamp": { "type": "string" }
        }
      },
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "object" },
          "error": { "type": ["string", "null"] },
          "timestamp": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "LevelUpRequest": {
        "type": "object",
        "required": ["exp_crystals"],
        "properties": {
          "exp_crystals": { "type": "integer", "minimum": 1 }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "authToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "authToken": { "type": "string" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "level": { "type": "integer" },
          "exp": { "type": "integer" },
          "crystals": { "type": "integer" },
          "gold": { "type": "integer" },
          "energy": { "type": "integer" },
          "max_energy": { "type": "integer" },
          "last_login": { "type": "string" }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "required": ["username", "email"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" }
        }
      }
    }
  }
}